+++ src/main/java/store/sonyk9919/api/domain/building/controller/BuildingLayoutController.java
@@ -3,7 +3,6 @@ package store.sonyk9919.api.domain.building.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,7 +15,6 @@ import store.sonyk9919.api.domain.auth.entity.AuthMember;
 import store.sonyk9919.api.domain.building.dto.BuildingLayoutDto;
 import store.sonyk9919.api.domain.building.dto.BuildingResponseDto;
 import store.sonyk9919.api.domain.building.service.BuildingLayoutService;
-import store.sonyk9919.api.domain.slot.dto.SlotMoveDto;
 
 @RestController
 @RequiredArgsConstructor
@@ -60,20 +58,5 @@ public class BuildingLayoutController {
             @PathVariable Integer slotNumber
     ) {
         return layoutService.demolishOf(authMember.getId(), slotNumber);
-    }
-
-    @Operation(
-            summary = "건물 이동 및 스왑",
-            description = "특정 슬롯에 있는 건물을 다른 슬롯으로 이동합니다."
-    )
-    @ApiResponses({
-            @ApiResponse(responseCode = "200", description = "이동 성공"),
-            @ApiResponse(responseCode = "400", description = "출발 슬롯에 건물이 없음 / 비활성화된 슬롯으로 이동 시도 / 동일한 슬롯 번호"),
-            @ApiResponse(responseCode = "404", description = "슬롯 없음"),
-            @ApiResponse(responseCode = "409", description = "동시성 충돌 (락 획득 실패)")
-    })
-    @PostMapping("/move")
-    public void move(@AuthMember AuthMemberDto authMember, @Valid @RequestBody SlotMoveDto slotMoveDto) {
-        layoutService.moveOf(authMember.getId(), slotMoveDto);
     }
 }
