+++ src/main/java/store/sonyk9919/api/domain/building/service/BuildingLayoutService.java
@@ -17,10 +17,8 @@ import store.sonyk9919.api.domain.island.entity.MemberIsland;
 import store.sonyk9919.api.domain.island.entity.ResourceType;
 import store.sonyk9919.api.domain.island.exception.IslandStatus;
 import store.sonyk9919.api.domain.island.service.ResourceService;
-import store.sonyk9919.api.domain.slot.dto.SlotMoveDto;
 import store.sonyk9919.api.domain.slot.dto.SlotResponseDto;
 import store.sonyk9919.api.domain.slot.entity.Slot;
-import store.sonyk9919.api.domain.slot.exception.SlotStatus;
 import store.sonyk9919.api.domain.slot.service.SlotQueryHelper;
 import store.sonyk9919.api.global.common.exception.CustomException;
 import store.sonyk9919.api.global.common.lock.DistributedLock;
@@ -104,24 +102,5 @@ public class BuildingLayoutService {
         if (yield.getRefundGem() > 0) {
             resourceService.add(island, ResourceType.GEM, yield.getRefundGem());
         }
-    }
-
-    @DistributedLock(keys = {
-            "'slot:' + #memberId + ':' + #slotMoveDto.fromSlotNumber",
-            "'slot:' + #memberId + ':' + #slotMoveDto.toSlotNumber"
-    })
-    @Transactional
-    public void moveOf(Long memberId, SlotMoveDto slotMoveDto) {
-        Integer fromSlotNumber = slotMoveDto.getFromSlotNumber();
-        Integer toSlotNumber = slotMoveDto.getToSlotNumber();
-
-        if (fromSlotNumber.equals(toSlotNumber)) throw new CustomException(SlotStatus.INVALID_MOVE);
-
-        Slot fromSlot = slotQueryHelper.getSlot(memberId, fromSlotNumber);
-        Slot toSlot = slotQueryHelper.getSlot(memberId, toSlotNumber);
-
-        if (!fromSlot.hasBuilding()) throw new CustomException(SlotStatus.SLOT_EMPTY);
-
-        fromSlot.swapBuildingWith(toSlot);
     }
 }
